Mobile Distro: BE - refactoring (own folder for Mobile Distro crud actions)
