+++ back/boxtribute_server/graph_ql/resolvers.py
@@ -31,17 +31,19 @@ from ..box_transfer.shipment import (
     update_shipment,
 )
 from ..enums import HumanGender, LocationType, TaggableObjectType, TransferAgreementType
-from ..models.crud import (
+from ..mobile_distribution.crud import (
     add_packing_list_entry_to_distribution_event,
     change_distribution_event_state,
-    create_beneficiary,
-    create_box,
     create_distribution_event,
     create_distribution_spot,
-    create_qr_code,
     delete_packing_list_entry,
     move_box_to_distribution_event,
     move_items_from_box_to_distribution_event,
+)
+from ..models.crud import (
+    create_beneficiary,
+    create_box,
+    create_qr_code,
     update_beneficiary,
     update_box,
 )
